Remove Xoodyak support
Xoodyak uses the Cyclist mode of operation, which, unlike Keccak's
sponge construction, includes the boundaries of chunks passed to
its equivalent to Update in the conceptual hash input.  As a
result, it may be difficult to drop into existing implementations
of existing protocols, and is unlikely to be useful to the
surrounding application.